refactor: include app data, settings in migrations
