+++ .github/workflows/test.yml
@@ -31,7 +31,7 @@ env:
   ETHERPAD_URL: 'http://etherpad:9001'
   LIBRARY_CLIENT_HOST: 'http://localhost:3113'
   FILE_STORAGE_ROOT_PATH: /
-  H5P_FILE_STORAGE_TYPE: file
+  H5P_FILE_STORAGE_TYPE: local
   H5P_PATH_PREFIX: h5p-content/
   H5P_STORAGE_ROOT_PATH: /tmp/h5p
   H5P_FILE_STORAGE_HOST: http://localhost:1081
@@ -47,7 +47,7 @@ env:
   REFRESH_TOKEN_JWT_SECRET: ${{ vars.MOCK_REFRESH_TOKEN_JWT_SECRET }}
   S3_FILE_ITEM_ACCESS_KEY_ID: graasp-user
   S3_FILE_ITEM_BUCKET: graasp
-  S3_FILE_ITEM_PLUGIN: true
+  FILE_STORAGE_TYPE: s3
   S3_FILE_ITEM_REGION: us-east-1
   S3_FILE_ITEM_SECRET_ACCESS_KEY: graasp-pwd
   SECURE_SESSION_SECRET_KEY: ${{ vars.MOCK_SECURE_SESSION_SECRET_KEY }}
