+++ src/services/file/repositories/s3.ts
@@ -18,9 +18,10 @@ import { FastifyBaseLogger } from 'fastify';
 
 import { UUID } from '@graasp/sdk';
 
-import { FileStorage, FileStorageType, S3_FILE_ITEM_HOST, TMP_FOLDER } from '../../../utils/config';
+import { S3_FILE_ITEM_HOST, TMP_FOLDER } from '../../../utils/config';
 import { S3FileConfiguration } from '../interfaces/configuration';
 import { FileRepository } from '../interfaces/fileRepository';
+import { FileStorageType } from '../types';
 import { S3_PRESIGNED_EXPIRATION } from '../utils/constants';
 import {
   DownloadFileUnexpectedError,
@@ -35,7 +36,6 @@ export class S3FileRepository implements FileRepository {
 
   constructor(options: S3FileConfiguration) {
     this.options = options;
-    this.fileStorageType = FileStorage.S3;
 
     const {
       s3Region: region,
